fix: use positive-lookaheads to ensure unquoted values are fully captured
I'm not 100% sure _why_ it worked before with a negative lookahead, but using
a positive-lookahead when capturing-until-delimiter feels a lot safer to me.

This patch also contains a minor refactor to how values are captured in the
block given to `String#scan(regexp, &block)` to splat-capture, using the
first non-nil value; performance tests indicate that this is a tiny bit better
than capturing each individually, but the primary motivation was to decouple
the method from the Regexp's exact implementation (allowing us in future to
add or remove more value-captures without changing the block signature).